services/spawn_service/app/requirements.txt
alembic==1.13.1
fastapi==0.110.2
httpx==0.27.0
psycopg[binary,pool]==3.1.19
pyyaml==6.0.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.30
uvicorn==0.29.0
